Move files in Components
